requirements.txt
psycopg2~=2.8.6
Flask~=1.1.2
httplib2~=0.19.0
google-api-python-client~=1.12.8
